+++ src/main/java/com/github/nthykier/debpkg/deb822/psi/impl/AbstractDeb822Paragraph.java
@@ -1,11 +1,18 @@
 package com.github.nthykier.debpkg.deb822.psi.impl;
 
 import com.github.nthykier.debpkg.deb822.Deb822LanguageSupport;
+import com.github.nthykier.debpkg.deb822.field.Deb822KnownField;
 import com.github.nthykier.debpkg.deb822.field.KnownFieldTable;
+import com.github.nthykier.debpkg.deb822.psi.Deb822ElementFactory;
 import com.github.nthykier.debpkg.deb822.psi.Deb822FieldValuePair;
 import com.github.nthykier.debpkg.deb822.psi.Deb822Paragraph;
+import com.github.nthykier.debpkg.deb822.psi.Deb822ValueParts;
 import com.intellij.extapi.psi.ASTWrapperPsiElement;
 import com.intellij.lang.ASTNode;
+import com.intellij.navigation.PsiElementNavigationItem;
+import com.intellij.openapi.util.NlsSafe;
+import com.intellij.psi.PsiElement;
+import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -13,7 +20,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 
-public abstract class AbstractDeb822Paragraph extends ASTWrapperPsiElement implements Deb822Paragraph {
+public abstract class AbstractDeb822Paragraph extends ASTWrapperPsiElement implements Deb822Paragraph, PsiElementNavigationItem {
 
   private Map<String, Deb822FieldValuePair> fieldMap;
   private Map<String, Deb822FieldValuePair> xPrefixedFieldMap;
@@ -75,5 +82,57 @@ public abstract class AbstractDeb822Paragraph extends ASTWrapperPsiElement imple
   public boolean isFirstParagraph() {
     Deb822Paragraph prevSibling = Deb822PsiImplUtil.getPreviousSiblingOfType(this.getPrevSibling(), Deb822Paragraph.class);
     return prevSibling == null;
+  }
+
+  @Override
+  public int getTextOffset() {
+    PsiElement psiElement = getNameIdentifier();
+    if (psiElement != null) {
+      return psiElement.getTextOffset();
+    }
+    return super.getTextOffset();
+  }
+
+  @Override
+  public @Nullable PsiElement getNameIdentifier() {
+    KnownFieldTable fieldTable = Deb822LanguageSupport.fromPsiElement(this).getKnownFieldTable();
+    Deb822KnownField knownField = fieldTable.getParagraphNamingField(this);
+    if (knownField == null) {
+      return null;
+    }
+    Deb822FieldValuePair pair = this.getFieldValuePair(knownField.getCanonicalFieldName());
+    if (pair == null) {
+      return null;
+    }
+    return pair.getValueParts();
+  }
+
+  @Override
+  public @Nullable String getName() {
+    PsiElement element = getNameIdentifier();
+    if (element == null) {
+      return null;
+    }
+    return element.getText();
+  }
+
+  @Override
+  public PsiElement setName(@NotNull String name) throws IncorrectOperationException {
+    Deb822ValueParts valueParts = Deb822ElementFactory.createValuePartsFromText(getProject(), "Foo: " + name);
+    Deb822ValueParts original = (Deb822ValueParts)this.getNameIdentifier();
+    if (original == null) {
+      throw new IncorrectOperationException("Unnamed");
+    }
+    original.replace(valueParts);
+    return this;
+  }
+
+  @Override
+  public @Nullable PsiElement getTargetElement() {
+    PsiElement name = getNameIdentifier();
+    if (name != null) {
+      return name;
+    }
+    return this;
   }
 }