+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ## [Unreleased]
 ### Added
 - Brace matching support inside the dependency language.
+- Add refactor support to rename binary packages.  It will attempt to rename relevant files in debian as
+  well (`oldname.foo` -> `newname.foo`).
 
 ### Changed
 - Bump lombok dependency fully to 1.18.22.  Previously 1.18.20 was still used in some cases.